Pulled out common test helper into mod.rs
