+++ src/model/quantity.rs
@@ -31,20 +31,9 @@ impl minijinja::value::Object for Quantity {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use cooklang::{Converter, CooklangParser, Extensions, ScaledRecipe};
-    use minijinja::{Environment, Value, context};
+    use crate::model::tests::get_recipe_and_env;
+    use minijinja::{Value, context};
     use test_case::test_case;
-
-    fn get_recipe_and_env<'a>(recipe: &str, template: &'a str) -> (ScaledRecipe, Environment<'a>) {
-        let recipe_parser = CooklangParser::new(Extensions::all(), Converter::default());
-        let (unscaled_recipe, _warnings) = recipe_parser.parse(recipe).into_result().unwrap();
-        let recipe = unscaled_recipe.scale(1.into(), &Converter::default());
-
-        let mut env: Environment<'a> = Environment::new();
-        env.add_template("test", template).unwrap();
-
-        (recipe, env)
-    }
 
     #[test_case("Crack @egg{1} into pan.", "{{ quantity }}", "1"; "number without unit")]
     #[test_case("Pour @flour{100%g} into bowl.", "{{ quantity }}", "100 g"; "number with unit")]
